backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
av==13.1.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloup==3.0.7
colorama==0.4.6
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.6.0
glcontext==3.0.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isosurfaces==0.1.2
Jinja2==3.1.6
jiter==0.10.0
jose==1.0.0
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.8.0
moderngl==5.12.0
moderngl-window==3.1.1
multidict==6.4.4
networkx==3.4.2
numpy==2.2.6
openai==1.82.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pyasn1==0.6.1
pycairo==1.28.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.4.3
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
scipy==1.15.3
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
svgelements==1.9.6
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
